backend/requirements.txt
fastapi
uvicorn
gunicorn
aiofiles
apscheduler
python-dotenv
requests
sqlalchemy
psycopg2-binary
python-telegram-bot
